move main.js, articles.js and articles.html
