+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "news-explorer-frontend",
-  "version": "1.32.0",
+  "version": "1.33.0",
   "description": "Yandex.Praktikum diploma 07-stream",
   "scripts": {
     "build": "rimraf dist && cross-env NODE_ENV=production webpack --mode production",
