+++ src/articles/articles.html
@@ -8,7 +8,7 @@
     <meta name="author" content="Sudomoin Pavel">
     <title>NewsExplorer</title>
     <link rel="stylesheet" href="<%= htmlWebpackPlugin.files.chunks.articles.css %>">
-    <link rel="shortcut icon" href="../<%=require('./images/favicon.ico').default%>" type="image/x-icon">
+    <link rel="shortcut icon" href="../<%=require('../images/favicon.ico').default%>" type="image/x-icon">
   </head>
   <body>
     <div class="page">
@@ -57,12 +57,12 @@
           <ul class="footer__soc-items">
             <li class="footer__soc-item">
               <a class="footer__soc-link link" target="_blank" href="https://github.com/pavel-sudomoin/news-explorer-frontend">
-                <img class="footer__soc-image" src="../<%=require('./images/github-logo.svg').default%>" alt="GitHub">
+                <img class="footer__soc-image" src="../<%=require('../images/github-logo.svg').default%>" alt="GitHub">
               </a>
             </li>
             <li class="footer__soc-item">
               <a class="footer__soc-link link" target="_blank" href="https://facebook.com">
-                <img class="footer__soc-image" src="../<%=require('./images/facebook-logo.svg').default%>" alt="Facebook">
+                <img class="footer__soc-image" src="../<%=require('../images/facebook-logo.svg').default%>" alt="Facebook">
               </a>
             </li>
           </ul>
