+++ src/articles/articles.js
@@ -1,34 +1,34 @@
 import '../pages/articles.css';
 
-import MainApi from './api/main-api';
+import MainApi from '../js/api/main-api';
 
 import Header from '../blocks/header/header';
 import Info from '../blocks/info/info';
 import NewsCardList from '../blocks/news-card-list/news-card-list';
 
-import headerRefresh from './utils/header-refresh';
-import infoRefresh from './utils/info-refresh';
-import savedArticlesRefresh from './utils/saved-articles-refresh';
-import getUserData from './utils/get-user-data';
-import getSavedArticlesData from './utils/get-saved-articles-data';
-import getUnathServerData from './utils/get-unath-server-data';
-import findArticleByEventTarget from './utils/find-article-by-event-target';
-import findArticleIndexById from './utils/find-article-index-by-id';
+import headerRefresh from '../js/utils/header-refresh';
+import infoRefresh from '../js/utils/info-refresh';
+import savedArticlesRefresh from '../js/utils/saved-articles-refresh';
+import getUserData from '../js/utils/get-user-data';
+import getSavedArticlesData from '../js/utils/get-saved-articles-data';
+import getUnathServerData from '../js/utils/get-unath-server-data';
+import findArticleByEventTarget from '../js/utils/find-article-by-event-target';
+import findArticleIndexById from '../js/utils/find-article-index-by-id';
 
-import { RESULT_SAVEDPAGE_CONTENT } from './constants/templates';
+import { RESULT_SAVEDPAGE_CONTENT } from '../js/constants/templates';
 import {
   HEADER_CONTAINER,
   RESULT_CONTAINER,
   INFO_CONTAINER,
-} from './constants/elements';
+} from '../js/constants/elements';
 import {
   MAIN_API_URL,
-} from './constants/api-config';
+} from '../js/constants/api-config';
 import {
   CANNOT_DELETE_ARTICLE_MESSAGE,
   CANNOT_LOGOUT_MESSAGE,
   CONFIRM_LOGOUT_MESSAGE,
-} from './constants/messages';
+} from '../js/constants/messages';
 
 
 // создаём инстансы классов
